requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0

# OpenAI and embeddings
openai>=1.25.0
tiktoken>=0.5.0

# Vector storage
pinecone-client>=3.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Document processing
python-docx>=1.0.0
openpyxl>=3.1.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Optional: For production deployment
gunicorn>=21.0.0
uvicorn>=0.24.0